requirements.txt
kfp==2.16.0
feast==0.61.0
mlflow==3.10.1
xgboost==2.1.1
scikit-learn==1.5.1
pandas==2.1.4
kafka-python==2.0.2
prometheus-client==0.14.1
fastapi==0.115.0
uvicorn==0.30.0
redis==5.0.1
numpy==1.26.4
google-cloud-storage==2.18.0
python-dotenv==1.0.0
httpx==0.27.0
pytest==8.3.2